liveSimpley/requirements.txt
django-webpack-loader
djangorestframework
django-cors-headers
django-rest-knox
pip-lock
coverage
